Printa 1 a 100, Foo, Baa, FooBaa
O exercício:
'''
Write a C/C++ or Java program that prints each number from 1 to 100 on a new line.

For each multiple of 3, print "Foo" instead of the number.
For each multiple of 5, print "Baa" instead of the number.
For multiple numbers simultaneously of 3 and 5, print "FooBaa", instead of the number.
'''
Foi resolvido. Para teste, basta chamar a função print_one_hundred_s1.